requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.13.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.10.0
apache-airflow-providers-common-compat==1.1.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.2
apache-airflow-providers-standard==1.3.0
apispec==6.6.1
argcomplete==3.5.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backports.strenum==1.3.1
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.8.2
cachelib==0.9.0
cadwyn==5.4.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.8.2
comm==0.2.3
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.0
cron-descriptor==1.4.3
croniter==3.0.3
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docutils==0.16
email_validator==2.2.0
entrypoints==0.4
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.1
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.2
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
graphviz==0.21
great-expectations==0.18.12
greenlet==3.0.3
grpcio==1.65.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.30.1
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==0.10.0
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lark==1.2.2
lazy-object-proxy==1.10.0
libcst==1.8.4
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
makefun==1.16.0
Mako==1.3.5
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.4
more-itertools==10.4.0
msgspec==0.19.0
multidict==6.0.5
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
ordered-set==4.1.0
overrides==7.7.0
packaging==24.1
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==11.3.0
pip==23.0.1
platformdirs==4.4.0
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==4.25.4
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pure-sasl==0.6.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
PyHive==0.7.0
PyJWT==2.9.0
pyparsing==3.2.3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.35.1
reportlab==4.4.3
requests==2.32.3
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.7.1
rich-argparse==1.5.2
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.20.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.12
scipy==1.13.1
Send2Trash==1.8.3
sentry-sdk==2.37.1
setproctitle==1.3.3
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.53
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.47.3
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
thrift==0.22.0
thrift-sasl==0.4.3
time-machine==2.15.0
tinycss2==1.4.0
tomli==2.2.1
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.4
types-python-dateutil==2.9.0.20250822
types-requests==2.32.4.20250809
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2024.1
tzlocal==5.3.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
uri-template==1.3.0
urllib3==1.26.19
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.2.3
widgetsnbextension==4.0.14
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.20.0
